started x86-64 arch support

+++ src/arch/includes/locks.h
@@ -4,7 +4,7 @@
 
 
 //bullshit code just for stuff to work
-void acquire_lock(uint32_t *addr, uint32_t bit);
-int try_acquire_lock(uint32_t *addr, uint32_t bit);
+void acquire_lock(uintptr_t *addr, uint8_t bit);
+int try_acquire_lock(uintptr_t *addr, uint8_t bit);
 
-void release_lock(uint32_t *addr, uint32_t bit);
+void release_lock(uintptr_t *addr, uint8_t bit);
