+++ src/arch/x86/panic/panic.h
@@ -6,7 +6,7 @@
 
 #define MAX_STACK_TRACE_SIZE 16
 
-void _panic_handler(uintptr_t isr_index, uint32_t err_code, cpu_registers_t* regs, uint32_t* call_stack);
+void _panic_handler(uintptr_t isr_index, size_t err_code, cpu_registers_t* regs, uintptr_t* call_stack);
 void _manual_panic(const char* error, const char* info);
 
 #endif // CRASHHNDL_H
